+++ app/src/main/java/com/example/arhms/interfaces/IObserver.java
@@ -1,6 +1,6 @@
 package com.example.arhms.interfaces;
 
 public interface IObserver {
-    public void update(long temperature, int heartRate);
+    public void update(int temperature, int heartRate);
     public String getString();
 }
