+++ app/src/main/java/com/example/arhms/utils/observer/Temperature.java
@@ -3,10 +3,10 @@ package com.example.arhms.utils.observer;
 import com.example.arhms.interfaces.IObserver;
 
 public class Temperature implements IObserver {
-    private long value;
+    private int value;
 
     @Override
-    public void update(long temperature, int heartRate) {
+    public void update(int temperature, int heartRate) {
         this.value = temperature;
         // do nothing with heart rate
     }
